feat: Package circlog for Nix / NixOS
This commit adds a default.nix which relies on nixpkgs buildGoModule to
package circlog for Nix and NixOS. The derivation currently imports the
local set of nixpkgs for ease of installation. This does mean the user
must have a nixpkgs channel setup with that exact name.

If they do not they can install it by importing their nixpkgs install
and calling `pkgs.callPackage (import <path or URL to this repo>)`.

To install in the current nix shell or environment you can simply run
`nix-env -if .` while inside the repo.